Final Update to the Main Menu
Cleaned up and commented mostly everything for the main menu scene. Deleted not-used variables and hopefully fixed the only UI issue left for the scene. If not, we will have to manually pass by it during the demo. Also, went ahead and made sure we aren't getting excessive warnings for the scene as well, so the amount of warnings in the Unity Editor has dropped to a lower amount as well.