Tutorial Zone Work Begun
Added some cover decorations to the tutorial zone, gonna be fiddling with the decorations for the scene some more. Also went ahead and added a light to grunt model that I will be editing as I go for clarity's sake of the player knowing where the grunt is in some of the darker areas of the game. Also I have begun re-tooling, cleaning up, and commenting on all of my previous code work so that It is easy to read, understand, and use. I will continue more work on this after a quick nap.