Code Cleanup Almost Complete
Setup the framework for the footstep sounds, went back and commented all of my code for the level UI script. Also added a new script that will replace the tutorial script and the entrance/exit script. Will be finishing that up tommorrow, and continue implementing fixes, changes, and quality of life increases.