More work on cleaning up code and re-tooling cutscenes.
Code cleanup is complete, just need to play around with the cutscenes for the tutorial and when you enter and exit a level. Once that's finished I'm plopping in the design for Level 3. Finally once that's finished, I'm adding in the UI to help the player identify whether or not they have a trap or taser shot ready.