Footsteps are In
Added an animation cycle to the player that tracks their feet, with footsteps being played in the background. Planning on adding in sounds to these remaining things: Traps, and Doors. Also going to update the taser sound effect. Finally I also fixed an issue with the camera that caused it to pan over to 0,0,0 because it never got set in the beginning.