Script Cleanup Complete
Just need to push the new script to Level's 1 2 and 3, then we will be in business to test. Unfortunately some things are going to have to be pushed back like the UI element to display your traps and taser shots remaining, as well as the cutscene that shows you even have the taser, but at the very least all of my code has been re-looked over, and tested multiple times that it does indeed work.