Enemy System Changes Complete
Completion is in the range  of 85-90% as changes will be made retroactively after environmental interactions are added.
Fixed issues that arose around the Brute being not properly tracking and persuing the player.
All enemies now properly enter their distracted states. Enemies have the functionality to properly react to the taser and traps.
Working on a basic environmental interaction of just blocking off a section.