Taser fixed
Taser now works when shooting in the cardinal directions. Enemies take damage and are frozen by the taser (can even be killed if shot enough times or when used correctly in combination with traps)
Taser was previously hitting the invisible detection sphere of the enemies, this has been remedied.
Made the Brute slightly faster when charging.